Router refactoring
