+++ src/components/App/style.css
@@ -54,3 +54,18 @@
   --accent-light-color: var(--accent-light-color-dark);
   --disabled-color: var(--disabled-color-dark);
 }
+
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+html, body, #root, .App {
+  width: 100%;
+  height: 100%;
+}
+
+.App {
+  overflow: auto;
+}
