+++ src/components/Header/meta.js
@@ -1,8 +1,10 @@
+import { paths } from "../../utils/constants";
+
 export const navContent = {
-    LOGO: 'BLOCKCHAIN WALLETS',
+    LOGO: "BLOCKCHAIN WALLETS",
     NAVLINKS: [
-        {id: 0, name: 'Блог', route: '/'},
-        {id: 1, name: 'Кошельки', route: '/wallets'},
-        {id: 2, name: 'Настройки', route: '/settings'},
-    ]
+        { id: 0, name: "Блог", route: paths.HOME },
+        { id: 1, name: "Кошельки", route: paths.WALLETS },
+        { id: 2, name: "Настройки", route: paths.SETTINGS },
+    ],
 };
