+++ src/components/NavMenu/index.jsx
@@ -1,20 +1,31 @@
+import React from "react";
 import { Link } from "react-router-dom";
-import './style.css';
+import "./style.css";
 
 export default function NavMenu(props) {
     const { navLinks } = props;
-    return (
-        <div className="navigation-menu">
-            <ul>
-                {
+
+    //TODO navLinks to memo to reduce rerenders
+    const NavigationLinks = () => {
+        return (
+            <React.Fragment>
+                {navLinks &&
                     navLinks.map((link) => (
                         <li key={link.id}>
                             <Link className="link" to={link.route}>
                                 {link.name}
                             </Link>
                         </li>
-                ))}
+                    ))}
+            </React.Fragment>
+        );
+    };
+
+    return (
+        <div className="navigation-menu">
+            <ul>
+                <NavigationLinks />
             </ul>
         </div>
     );
-};
+}
